+++ src/routes/login/resetPassword/+page.server.ts
@@ -2,7 +2,7 @@ import { SERVER_ERROR } from '$lib/helperTypes';
 import { PasswordResetRepo } from '$lib/passwordReset';
 import { prisma } from '$lib/prisma';
 import { auth } from '$lib/server/lucia';
-import { resetPasswordSchema } from '$lib/user';
+import { resetPasswordSchema } from '$lib/zodSchemas';
 import { error, fail, redirect } from '@sveltejs/kit';
 import { message, superValidate } from 'sveltekit-superforms/server';
 import type { Actions, PageServerLoad } from './$types';
